requirements.txt
streamlit>=1.28.0
SpeechRecognition>=3.10.0
soundfile>=0.12.1
pandas>=2.0.0
pydub>=0.25.1
numpy>=1.24.0 
google-generativeai
